put PrimaryLepton into own file to make code clearer
